backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.50.0
langchain-openai>=0.2.0
langchain-postgres>=0.0.12
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
slowapi>=0.1.9
psycopg[binary]>=3.1.0
sentence-transformers>=3.0.0
